Refactor jitutils source code
Created a new `src/util` directory to put code common to all tools.
Created a util.cs there with a "Utility" class. Moved several
functions there that were already copied across various tools.

In jit-diff, created new diff.cs, install.cs, and uninstall.cs files
to split out the code specific to those commands into their own files.
