+++ src/jit-diff/jit-diff.cs
@@ -3,7 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Diagnostics;
+//using System.Diagnostics;
 using System.CommandLine;
 using System.IO;
 using System.Collections.Generic;
@@ -11,27 +11,14 @@ using System.Text.RegularExpressions;
 using System.Linq;
 using Microsoft.DotNet.Cli.Utils;
 using Microsoft.DotNet.Tools.Common;
-using Newtonsoft.Json;
+//using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
-using System.Collections.Concurrent;
+//using System.Collections.Concurrent;
 using System.Threading.Tasks;
 using System.Runtime.InteropServices;
 
 namespace ManagedCodeGen
 {
-    public class Utility
-    {
-        public static string CombinePath(string basePath, string[] pathComponents)
-        {
-            string resultPath = basePath;
-            foreach (var dir in pathComponents)
-            {
-                resultPath = Path.Combine(resultPath, dir);
-            }
-            return resultPath;
-        }
-    }
-
     public class AssemblyInfo
     {
         // Contains full path to assembly.
@@ -42,7 +29,7 @@ namespace ManagedCodeGen
         public string OutputPath { get; set; }
     }
 
-    public class jitdiff
+    public partial class jitdiff
     {
         // Supported commands.  List to view information from the CI system, and Copy to download artifacts.
         public enum Commands
@@ -224,7 +211,7 @@ namespace ManagedCodeGen
             {
                 // Extract system RID from dotnet cli
                 List<string> commandArgs = new List<string> { "--info" };
-                CommandResult result = TryCommand("dotnet", commandArgs, true);
+                CommandResult result = Utility.TryCommand("dotnet", commandArgs, true);
 
                 if (result.ExitCode != 0)
                 {
@@ -303,7 +290,7 @@ namespace ManagedCodeGen
 
                 if (needDiffRoot)
                 {
-                    _diffRoot = GetRepoRoot();
+                    _diffRoot = Utility.GetRepoRoot(Verbose);
                 }
 
                 if (needOutputPath && (_diffRoot != null))
@@ -557,54 +544,6 @@ namespace ManagedCodeGen
                     Console.WriteLine("--diff: {0}", _diffPath);
                     Console.WriteLine("--test_root: {0}", _testPath);
                 }
-            }
-
-            private string GetRepoRoot()
-            {
-                // git rev-parse --show-toplevel
-                List<string> commandArgs = new List<string> { "rev-parse", "--show-toplevel" };
-                CommandResult result = TryCommand("git", commandArgs, true);
-                if (result.ExitCode != 0)
-                {
-                    if (Verbose)
-                    {
-                        Console.Error.WriteLine("'git rev-parse --show-toplevel' returned non-zero ({0})", result.ExitCode);
-                    }
-                    return null;
-                }
-
-                var lines = result.StdOut.Split(new[] { Environment.NewLine }, StringSplitOptions.None);
-                var git_root = lines[0];
-                var repo_root = git_root.Replace('/', Path.DirectorySeparatorChar);
-
-                // Is it actually the dotnet/coreclr repo?
-                commandArgs = new List<string> { "remote", "-v" };
-                result = TryCommand("git", commandArgs, true);
-                if (result.ExitCode != 0)
-                {
-                    if (Verbose)
-                    {
-                        Console.Error.WriteLine("'git remote -v' returned non-zero ({0})", result.ExitCode);
-                    }
-                    return null;
-                }
-
-                bool isCoreClr = result.StdOut.Contains("/coreclr");
-                if (!isCoreClr)
-                {
-                    if (Verbose)
-                    {
-                        Console.Error.WriteLine("Doesn't appear to be the dotnet/coreclr repo:");
-                        Console.Error.WriteLine(result.StdOut);
-                    }
-                    return null;
-                }
-
-                if (Verbose)
-                {
-                    Console.WriteLine("Repo root: " + repo_root);
-                }
-                return repo_root;
             }
 
             private void DeriveOutputTag()
@@ -1350,607 +1289,6 @@ namespace ManagedCodeGen
             }
 
             return ret;
-        }
-
-        class ScriptResolverPolicyWrapper : ICommandResolverPolicy
-        {
-            public CompositeCommandResolver CreateCommandResolver() => ScriptCommandResolverPolicy.Create();
-        }
-
-        public static CommandResult TryCommand(string name, IEnumerable<string> commandArgs, bool capture = false)
-        {
-            try
-            {
-                Command command = Command.Create(new ScriptResolverPolicyWrapper(), name, commandArgs);
-
-                if (capture)
-                {
-                    // Capture stdout/stderr for consumption within tool.
-                    command.CaptureStdOut();
-                    command.CaptureStdErr();
-                }
-                else
-                {
-                    // Wireup stdout/stderr so we can see output.
-                    command.ForwardStdOut();
-                    command.ForwardStdErr();
-                }
-
-                return command.Execute();
-            }
-            catch (CommandUnknownException e)
-            {
-                Console.Error.WriteLine("\nerror: {0} command not found!  Add {0} to the path.", name, e);
-                Environment.Exit(-1);
-                return CommandResult.Empty;
-            }
-        }
-
-        public static int InstallCommand(Config config)
-        {
-            var configFilePath = Path.Combine(config.JitUtilsRoot, s_configFileName);
-            string configJson = File.ReadAllText(configFilePath);
-            var jObj = JObject.Parse(configJson);
-
-            if ((jObj[s_configFileRootKey] == null) || (jObj[s_configFileRootKey]["tools"] == null))
-            {
-                Console.Error.WriteLine("\"install\" doesn't know how to add the \"" + s_configFileRootKey + "\":\"tools\" section to the config file");
-                return -1;
-            }
-
-            if ((config.PlatformMoniker == null) || (GetBuildOS(config.PlatformMoniker) == null))
-            {
-                return -1;
-            }
-
-            var tools = (JArray)jObj[s_configFileRootKey]["tools"];
-
-            // Early out if the tool is already installed. We can only do this if we're not doing
-            // "--last_successful", in which case we don't know what the build number (and hence
-            // tag) is.
-            string tag = null;
-            if (!config.DoLastSucessful)
-            {
-                tag = String.Format("{0}-{1}", config.JobName, config.Number);
-                if (tools.Where(x => (string)x["tag"] == tag).Any())
-                {
-                    Console.Error.WriteLine("{0} is already installed in the " + s_configFileName + ". Remove before re-install.", tag);
-                    return -1;
-                }
-            }
-
-            string toolPath = Path.Combine(config.JitUtilsRoot, "tools");
-
-            // Issue cijobs command to download bits            
-            List<string> cijobsArgs = new List<string>();
-
-            cijobsArgs.Add("copy");
-
-            cijobsArgs.Add("--job");
-            cijobsArgs.Add(config.JobName);
-
-            if (config.BranchName != null)
-            {
-                cijobsArgs.Add("--branch");
-                cijobsArgs.Add(config.BranchName);
-            }
-
-            if (config.DoLastSucessful)
-            {
-                cijobsArgs.Add("--last_successful");
-            }
-            else
-            {
-                cijobsArgs.Add("--number");
-                cijobsArgs.Add(config.Number);
-            }
-
-            cijobsArgs.Add("--unzip");
-
-            cijobsArgs.Add("--output");
-            cijobsArgs.Add(toolPath);
-
-            if (config.Verbose)
-            {
-                Console.WriteLine("Command: {0} {1}", "cijobs", String.Join(" ", cijobsArgs));
-            }
-
-            CommandResult result = TryCommand("cijobs", cijobsArgs);
-
-            if (result.ExitCode != 0)
-            {
-                Console.Error.WriteLine("cijobs command returned with {0} failures", result.ExitCode);
-                return result.ExitCode;
-            }
-
-            // There is a convention that cijobs creates a directory to store the job within
-            // the toolPath named:
-            //      <job-name>-<version-number>
-            // for example:
-            //      checked_windows_nt-1234
-            //
-            // However, if we passed "--last_successful", we don't know that number! So, figure it out.
-
-            if (config.DoLastSucessful)
-            {
-                // Find the largest numbered build with this job name.
-                int maxBuildNum = -1;
-                foreach (var dir in Directory.EnumerateDirectories(toolPath))
-                {
-                    Regex dirPattern = new Regex(@"(.*)-(.*)");
-                    Match dirMatch = dirPattern.Match(dir);
-                    if (dirMatch.Success)
-                    {
-                        if (int.TryParse(dirMatch.Groups[1].Value, out int thisBuildNum))
-                        {
-                            if (thisBuildNum > maxBuildNum)
-                            {
-                                maxBuildNum = thisBuildNum;
-                            }
-                        }
-                    }
-                }
-
-                if (maxBuildNum == -1)
-                {
-                    Console.Error.WriteLine("Error: couldn't determine last successful build directory in {0}", toolPath);
-                    return -1;
-                }
-
-                string buildNum = maxBuildNum.ToString();
-                tag = String.Format("{0}-{1}", config.JobName, buildNum);
-            }
-
-            toolPath = Path.Combine(toolPath, tag);
-
-            JObject newTool = new JObject();
-            newTool.Add("tag", tag);
-            string platformPath = Path.Combine(toolPath, "Product");
-            if (!Directory.Exists(platformPath))
-            {
-                Console.Error.WriteLine("cijobs didn't create or populate directory {0}", platformPath);
-                return 1;
-            }
-
-            string buildOS = GetBuildOS(config.PlatformMoniker).ToUpper();
-            foreach (var dir in Directory.EnumerateDirectories(platformPath))
-            {
-                if (Path.GetFileName(dir).ToUpper().Contains(buildOS))
-                {
-                    newTool.Add("path", Path.GetFullPath(dir));
-                    if (tools.HasValues)
-                    {
-                        tools.Last.AddAfterSelf(newTool);
-                    }
-                    else
-                    {
-                        tools.Add(newTool);
-                    }
-                    break;
-                }
-            }
-
-            // Overwrite current config.json with new data.
-            using (var file = File.CreateText(configFilePath))
-            {
-                using (JsonTextWriter writer = new JsonTextWriter(file))
-                {
-                    writer.Formatting = Formatting.Indented;
-                    jObj.WriteTo(writer);
-                }
-            }
-
-            return 0;
-        }
-
-        public static int UninstallCommand(Config config)
-        {
-            var configFilePath = Path.Combine(config.JitUtilsRoot, s_configFileName);
-            string configJson = File.ReadAllText(configFilePath);
-            var jObj = JObject.Parse(configJson);
-
-            if ((jObj[s_configFileRootKey] == null) || (jObj[s_configFileRootKey]["tools"] == null))
-            {
-                Console.Error.WriteLine("Error: no \"" + s_configFileRootKey + "\":\"tools\" section in the config file");
-                return -1;
-            }
-
-            var tools = (JArray)jObj[s_configFileRootKey]["tools"];
-            var elem = tools.Children()
-                            .Where(x => (string)x["tag"] == config.Tag);
-            if (!elem.Any())
-            {
-                Console.WriteLine("{0} is not installed in {1}.", config.Tag, s_configFileName);
-                return -1;
-            }
-
-            var jobj = elem.First();
-            string path = (string)jobj["path"];
-            if (path != null)
-            {
-                Console.WriteLine("Warning: you should remove install directory {0}.", path);
-
-                // We could do this:
-                //      Directory.Delete(path, true);
-                // However, the "install" command copies down a lot more than just this directory,
-                // so removing this directory still leaves around a lot of stuff.
-            }
-
-            Console.WriteLine("Removing tag {0} from config file.", config.Tag);
-            jobj.Remove();
-
-            // Overwrite current config.json with new data.
-            using (var file = File.CreateText(configFilePath))
-            {
-                using (JsonTextWriter writer = new JsonTextWriter(file))
-                {
-                    writer.Formatting = Formatting.Indented;
-                    jObj.WriteTo(writer);
-                }
-            }
-
-            return 0;
-        }
-
-        public class DiffTool
-        {
-            private class DasmWorkItem
-            {
-                public List<string> DasmArgs { get; set; }
-
-                public DasmWorkItem(List<string> dasmArgs)
-                {
-                    DasmArgs = dasmArgs;
-                }
-            }
-
-            private Config m_config;
-            private List<Task<int>> DasmWorkTasks = new List<Task<int>>();
-
-            public DiffTool(Config config)
-            {
-                m_config = config;
-            }
-
-            // Returns a count of the number of failures.
-            private int RunDasmTool(DasmWorkItem item)
-            {
-                int dasmFailures = 0;
-
-                string command = s_asmTool + " " + String.Join(" ", item.DasmArgs);
-                Console.WriteLine("Dasm command: {0}", command);
-                CommandResult result = TryCommand(s_asmTool, item.DasmArgs);
-                if (result.ExitCode != 0)
-                {
-                    Console.Error.WriteLine("Dasm command \"{0}\" returned with {1} failures", command, result.ExitCode);
-                    dasmFailures += result.ExitCode;
-                }
-
-                return dasmFailures;
-            }
-
-            private void StartDasmWork(List<string> args)
-            {
-                Task<int> task = Task<int>.Factory.StartNew(() => RunDasmTool(new DasmWorkItem(args)));
-                DasmWorkTasks.Add(task);
-
-                if (m_config.Verbose)
-                {
-                    string command = String.Join(" ", args);
-                    Console.WriteLine("Started dasm command \"{0}\"", command);
-                }
-
-                if (m_config.Sequential)
-                {
-                    Task.WaitAll(task);
-                }
-            }
-
-            private void StartDasmWorkOne(List<string> commandArgs, string tagBaseDiff, string clrPath, AssemblyInfo assemblyInfo)
-            {
-                List<string> args = ConstructArgs(commandArgs, clrPath);
-
-                string outputPath = Path.Combine(m_config.OutputPath, tagBaseDiff, assemblyInfo.OutputPath);
-                args.Add("--output");
-                args.Add(outputPath);
-
-                args.Add(assemblyInfo.Path);
-
-                StartDasmWork(args);
-            }
-
-            private void StartDasmWorkBaseDiff(List<string> commandArgs, List<AssemblyInfo> assemblyWorkList)
-            {
-                foreach (var assemblyInfo in assemblyWorkList)
-                {
-                    if (m_config.HasBasePath)
-                    {
-                        StartDasmWorkOne(commandArgs, "base", m_config.BasePath, assemblyInfo);
-                    }
-                    if (m_config.HasDiffPath)
-                    {
-                        StartDasmWorkOne(commandArgs, "diff", m_config.DiffPath, assemblyInfo);
-                    }
-                }
-            }
-
-            private List<string> ConstructArgs(List<string> commandArgs, string clrPath)
-            {
-                List<string> dasmArgs = commandArgs.ToList();
-                dasmArgs.Add("--crossgen");
-                if (m_config.HasCrossgenExe)
-                {
-                    dasmArgs.Add(m_config.CrossgenExe);
-
-                    var jitPath = Path.Combine(clrPath, GetJitLibraryName(m_config.PlatformMoniker));
-                    if (!File.Exists(jitPath))
-                    {
-                        Console.Error.WriteLine("clrjit not found at " + jitPath);
-                        return null;
-                    }
-
-                    dasmArgs.Add("--jit");
-                    dasmArgs.Add(jitPath);
-                }
-                else
-                {
-                    var crossgenPath = Path.Combine(clrPath, GetCrossgenExecutableName(m_config.PlatformMoniker));
-                    if (!Directory.Exists(crossgenPath))
-                    {
-                        Console.Error.WriteLine("crossgen not found at " + crossgenPath);
-                        return null;
-                    }
-
-                    dasmArgs.Add(crossgenPath);
-                }
-
-                return dasmArgs;
-            }
-
-            // Returns:
-            // 0 on success,
-            // Otherwise, a count of the number of failures generating asm, as reported by the asm tool.
-            private int RunDasmTool(List<string> commandArgs, List<AssemblyInfo> assemblyWorkList)
-            {
-                StartDasmWorkBaseDiff(commandArgs, assemblyWorkList);
-
-                int dasmFailures = 0;
-
-                try
-                {
-                    Task<int>[] taskArray = DasmWorkTasks.ToArray();
-                    Task.WaitAll(taskArray);
-
-                    foreach (Task<int> t in taskArray)
-                    {
-                        dasmFailures += t.Result;
-                    }
-                }
-                catch (AggregateException ex)
-                {
-                    Console.Error.WriteLine("Dasm task failed with {0}", ex.Message);
-                    dasmFailures += ex.InnerExceptions.Count;
-                }
-
-                if (dasmFailures != 0)
-                {
-                    Console.Error.WriteLine("Dasm commands returned with total of {0} failures", dasmFailures);
-                }
-
-                return dasmFailures;
-            }
-
-            // Returns 0 on success, 1 on failure.
-            public static int DiffCommand(Config config)
-            {
-                string diffString = "";
-
-                if (config.CoreLib)
-                {
-                    diffString += "System.Private.CoreLib.dll";
-                }
-                else if (config.DoFrameworks)
-                {
-                    diffString += "System.Private.CoreLib.dll, framework assemblies";
-                }
-
-                if (config.Benchmarks)
-                {
-                    if (!String.IsNullOrEmpty(diffString)) diffString += ", ";
-                    diffString += "benchstones and benchmarks game in " + config.TestRoot;
-                }
-                else if (config.DoTestTree)
-                {
-                    if (!String.IsNullOrEmpty(diffString)) diffString += ", ";
-                    diffString += "assemblies in " + config.TestRoot;
-                }
-
-                Console.WriteLine("Beginning diff of {0}", diffString);
-
-                // Create subjob that runs jit-dasm, which should be in path, with the 
-                // relevent coreclr assemblies/paths.
-
-                List<string> commandArgs = new List<string>();
-
-                commandArgs.Add("--platform");
-                commandArgs.Add(config.CoreRoot);
-
-                if (config.GenerateGCInfo)
-                {
-                    commandArgs.Add("--gcinfo");
-                }
-
-                if (config.Verbose)
-                {
-                    commandArgs.Add("--verbose");
-                }
-
-                List<AssemblyInfo> assemblyWorkList = GenerateAssemblyWorklist(config);
-
-                DiffTool diffTool = new DiffTool(config);
-                int dasmFailures = diffTool.RunDasmTool(commandArgs, assemblyWorkList);
-
-                // Analyze completed run.
-
-                if (config.DoAnalyze && config.HasDiffPath && config.HasBasePath)
-                {
-                    List<string> analysisArgs = new List<string>();
-
-                    analysisArgs.Add("--base");
-                    analysisArgs.Add(Path.Combine(config.OutputPath, "base"));
-                    analysisArgs.Add("--diff");
-                    analysisArgs.Add(Path.Combine(config.OutputPath, "diff"));
-                    analysisArgs.Add("--recursive");
-
-                    Console.WriteLine("Analyze command: {0} {1}",
-                        s_analysisTool, String.Join(" ", analysisArgs));
-
-                    CommandResult analyzeResult = TryCommand(s_analysisTool, analysisArgs);
-                }
-
-                // Report any failures to generate asm at the very end (again). This is so
-                // this information doesn't get buried in previous output.
-
-                if (dasmFailures != 0)
-                {
-                    Console.Error.WriteLine("");
-                    Console.Error.WriteLine("Warning: {0} failures detected generating asm", dasmFailures);
-
-                    return 1; // failure result
-                }
-                else
-                {
-                    return 0; // success result
-                }
-            }
-
-            public static List<AssemblyInfo> GenerateAssemblyWorklist(Config config)
-            {
-                List<AssemblyInfo> assemblyInfoList = new List<AssemblyInfo>();
-
-                // CoreLib and the frameworks add specific files, not directories.
-                // These files will all be put in the same output directory. This
-                // works because they all have unique names, and live in the same
-                // source directory already.
-                if (config.CoreLib || config.DoFrameworks)
-                {
-                    foreach (var assembly in config.CoreLib ? s_CoreLibAssembly : s_frameworkAssemblies)
-                    {
-                        string fullPathAssembly = Path.Combine(config.CoreRoot, assembly);
-
-                        if (!File.Exists(fullPathAssembly))
-                        {
-                            Console.Error.WriteLine("Warning: can't find framework assembly {0}", fullPathAssembly);
-                            continue;
-                        }
-
-                        AssemblyInfo info = new AssemblyInfo
-                        {
-                            Path = fullPathAssembly,
-                            OutputPath = ""
-                        };
-
-                        assemblyInfoList.Add(info);
-                    }
-                }
-
-                // The tests are in a tree hierarchy of directories. We will output these to a tree
-                // structure matching their source tree structure, to avoid name conflicts.
-                if (config.Benchmarks || config.DoTestTree)
-                {
-                    string basepath = config.Benchmarks ? Utility.CombinePath(config.TestRoot, s_benchmarksPath) : config.TestRoot;
-                    foreach (var dir in config.Benchmarks ? s_benchmarkDirectories : s_testDirectories)
-                    {
-                        string fullPathDir = Path.Combine(basepath, dir);
-
-                        if (!Directory.Exists(fullPathDir))
-                        {
-                            Console.Error.WriteLine("Warning: can't find test directory {0}", fullPathDir);
-                            continue;
-                        }
-
-                        // For the directory case create a stack and recursively find any
-                        // assemblies for compilation.
-                        List<AssemblyInfo> directoryAssemblyInfoList = IdentifyAssemblies(basepath, fullPathDir, config);
-
-                        // Add info generated at this directory
-                        assemblyInfoList.AddRange(directoryAssemblyInfoList);
-                    }
-                }
-
-                return assemblyInfoList;
-            }
-
-            // Check to see if the passed filePath is to an assembly.
-            private static bool IsAssembly(string filePath)
-            {
-                try
-                {
-                    System.Reflection.AssemblyName diffAssembly =
-                        System.Runtime.Loader.AssemblyLoadContext.GetAssemblyName(filePath);
-                }
-                catch (System.IO.FileNotFoundException)
-                {
-                    // File not found - not an assembly
-                    // TODO - should we log this case?
-                    return false;
-                }
-                catch (System.BadImageFormatException)
-                {
-                    // Explictly not an assembly.
-                    return false;
-                }
-                catch (System.IO.FileLoadException)
-                {
-                    // This is an assembly but it just happens to be loaded.
-                    // (leave true in so as not to rely on fallthrough)
-                    return true;
-                }
-
-                return true;
-            }
-
-            // Recursively search for assemblies from a path given by `rootPath`.
-            // `basePath` specifies the root directory for the purpose of constructing a relative output path.
-            //      It must be a prefix of `rootPath`.
-            private static List<AssemblyInfo> IdentifyAssemblies(string basePath, string rootPath, Config config)
-            {
-                List<AssemblyInfo> assemblyInfoList = new List<AssemblyInfo>();
-                string fullBasePath = Path.GetFullPath(basePath);
-                string fullRootPath = Path.GetFullPath(rootPath);
-
-                // Get files that could be assemblies, but discard currently ngen'd assemblies.
-                var subFiles = Directory.EnumerateFiles(fullRootPath, "*", SearchOption.AllDirectories)
-                    .Where(s => (s.EndsWith(".exe") || s.EndsWith(".dll")) && !s.Contains(".ni."));
-
-                foreach (var filePath in subFiles)
-                {
-                    if (config.Verbose)
-                    {
-                        Console.WriteLine("Scanning: {0}", filePath);
-                    }
-
-                    // skip if not an assembly
-                    if (!IsAssembly(filePath))
-                    {
-                        continue;
-                    }
-
-                    string directoryName = Path.GetDirectoryName(filePath);
-                    string fullDirectoryName = Path.GetFullPath(directoryName);
-                    string outputPath = fullDirectoryName.Substring(fullBasePath.Length).TrimStart(Path.DirectorySeparatorChar);
-
-                    AssemblyInfo info = new AssemblyInfo
-                    {
-                        Path = filePath,
-                        OutputPath = outputPath
-                    };
-
-                    assemblyInfoList.Add(info);
-                }
-
-                return assemblyInfoList;
-            }
         }
     }
 }
